pyproject.toml
[project]
name = "syntax-and-empathy"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
  "jupyter",
  "ipykernel",
  "pillow>=11.3.0",
]
